Attendance Partially Done

+++ server/src/main/java/com/postgresql/pgms/repo/CourseRepo.java
@@ -8,6 +8,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import java.util.List;
 
 @RepositoryRestResource
-public interface CourseRepo extends JpaRepository<course,Long> {
+public interface CourseRepo extends JpaRepository<course,Integer> {
 
 }
